Revert "Extract url sync in separate hook"
This reverts commit 3a9cc8567fcaa61d1b73640cfdd865914e6d7655.
